requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Vision and AI capabilities
langchain==0.1.5
openai==1.1.1
httpx==0.24.1
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# Geolocation and metadata extraction
exifread==3.0.0
geopy==2.4.1
folium==0.15.0

# Frontend
streamlit==1.30.0
plotly==5.18.0

# Utils
python-dotenv==1.0.0
requests==2.31.0 